django-amazon-s3-test/requirements.txt
Django==3.1.4
boto3==1.16.28
django-storages==1.10.1
